Update pylupdate5 to pylupdate6
Previously the pylupdate5 command was used to get strings from Python
code but this version have some flaws that make hard for developer add
translations support. Most of this problems are fixed in pylupdate6,
this version is preferred.

Now is possible to use `tr(string)` and the context will be assigned
automatically based on the class name letting the developer to have a
cleaner code.